[FIX] #S09P22A406-229 마이페이지 API 완성

+++ frontend/app/src/main/java/com/a406/horsebit/DataModels.kt
@@ -106,12 +106,12 @@ data class MyAssetModel(
 
 // 마이페이지 총자산 component
 data class MyTotalAssetModel(
-    val totalAsset : Int,
-    val cashBalance: Int,
-    val totalPurchase: Int,
-    val totalEvaluation: Int,
-    val profitOrLoss: Int,
-    val returnRate: Int,
+    val totalAsset : Double,
+    val cashBalance: Double,
+    val totalPurchase: Double,
+    val totalEvaluation: Double,
+    val profitOrLoss: Double,
+    val returnRate: Double,
 )
 
 // 로그인
